chore(ibc): refactor, rename /parse_ibc to /ibc
